used delta in update method

+++ StevenHuang.spritebuilder/GameplayManager.h
@@ -14,6 +14,6 @@
 
 @property (nonatomic) NSInteger level;
 @property (assign) bool paused;
-@property (assign) int roundCounter;
+@property (assign) float roundCounter;
 
 @end
